backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
SQLAlchemy>=2.0.0
pydantic-settings>=2.2.0
python-multipart>=0.0.9
Pillow>=10.0.0
